+++ src/com/oldofus/game/objects/dep/Creature.java
@@ -7,8 +7,8 @@ import com.oldofus.models.Spell.SpellStats;
 public abstract class Creature {
 
     protected Stats baseStats = new Stats();
-    protected int level;
-    protected int gfxID;
+    protected short level;
+    protected short gfxID;
     protected ArrayList<SpellStats> _spells;
     protected String[] colors = new String[3];
     protected String name;
@@ -45,10 +45,10 @@ public abstract class Creature {
      *
      * @return
      */
-    public Integer getInitiative() {
-        int fact = 4;
-        int pvmax = 100;
-        int pv = 100;
+    public Short getInitiative() {
+        short fact = 4;
+        short pvmax = 100;
+        short pv = 100;
         /*if (_classe == Constants.CLASS_SACRIEUR) {
          fact = 8;
          }*/
@@ -60,9 +60,9 @@ public abstract class Creature {
         coef += getTotalStats().get(Element.INTEL);
         coef += getTotalStats().get(Element.FORCE);
 
-        int init = 1;
+        short init = 1;
         if (pvmax != 0) {
-            init = (int) (coef * ((double) pv / (double) pvmax));
+            init = (short) (coef * ((double) pv / (double) pvmax));
         }
         if (init < 0) {
             init = 0;
@@ -70,11 +70,11 @@ public abstract class Creature {
         return init;
     }
 
-    public Integer getLevel() {
+    public Short getLevel() {
         return level;
     }
 
-    public Integer getGfxID() {
+    public Short getGfxID() {
         return gfxID;
     }
